testfix

+++ src/test/java/htw/webtech/WT_todo/TodoRepositoryTest.java
@@ -29,8 +29,8 @@ class TodoRepositoryTest {
         todoRepository.save(a);
         todoRepository.save(b);
 
-        long deleted = todoRepository.deleteByDoneTrue();
-        assertThat(deleted).isEqualTo(1L);
+        int deleted = todoRepository.deleteByDoneTrue();
+        assertThat(deleted).isEqualTo(1);
         assertThat(todoRepository.findAll()).hasSize(1);
         assertThat(todoRepository.findAll().get(0).isDone()).isFalse();
     }
